Move functions to inc/
